requirements.txt
paho-mqtt
bleak
Pillow
